+++ src/components/BreadCrumbItem/BreadCrumbItem.jsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import PropTypes from "prop-types";
-import "./breadCrumbItem.scss";
 
 const BreadcrumbItem = ({ to, text }) => (
    <>
