+++ src/components/BreadCrumbItem/breadCrumbItem.scss
@@ -1,5 +1,3 @@
-@import "../../styles/style.scss";
-
 .navigation__block-current {
    font-weight: 500;
    font-size: 16px;
