+++ src/components/Footer/footer.scss
