+++ src/components/TopProductItems/TopProductsItems.jsx
@@ -2,7 +2,7 @@ import React, { memo, useEffect, useState } from "react";
 import "../../styles/global.scss";
 import "./topProductItems.scss";
 import useServer from "../../hooks/useServer";
-import ProductCard from "../OOproductCard/ProductCard";
+import ProductCard from "../ProductCard/ProductCard";
 
 function TopProductItem() {
   const { getAllProducts } = useServer();
