+++ src/pages/Home/Home.jsx
@@ -1,9 +1,9 @@
 import Benefits from "../../components/OurBenefits/OurBenefitsList";
 import OurPartners from "../../components/OurPartners/OurPartners";
 import TopItemsSlider from "../../components/TopItemsSlider/TopItemsSlider";
-import TopProductItem from "../../components/OOtopProductItems/TopProductsItems";
+import TopProductItem from "../../components/TopProductItems/TopProductsItems";
 import content from "../../components/OurBenefits/content.json";
-import Breadcrumb from "../../components/OObreadCrumb/BreadCrumb";
+import Breadcrumb from "../../components/BreadCrumb/BreadCrumb";
 
 
 const Home = () => {
