+++ src/styles/style.scss
@@ -4,10 +4,11 @@
 @import "./mixins.scss";
 @import "../pages/LoginPage/loginPage.scss";
 @import "../pages/RegistrationPage/RegistrationPage.scss";
-@import "../components/OOproductCard/productCard.scss";
+@import "../components/ProductCard/productCard.scss";
 @import "../components/OurBenefits/ourBenefits.scss";
 @import "../components/OurPartners/ourPartners.scss";
 @import "../components/TopItemsSlider/topItemsSlider.scss";
 @import "../components/TopItemsSlider/slick.scss";
 @import "../components/TopItemsSlider/slick-theme.scss";
 @import "../components/ProductsSection/productsSection.scss";
+@import "../components/BreadcrumbItem/breadCrumbItem.scss";
